Clean up stale copy and duplication from Bots/Remote Control/IM Notify IA
Fixes surfaced by a self-review pass over the IA rework: leftover
"Overview"/"Notify"/"left nav"/"tab bar" wording from earlier iterations,
a duplicated PLATFORM_BRAND_ICONS map (now shared via platformGuides.tsx),
a duplicated per-platform active/total counter (now countBotsByPlatform in
types/bot.ts), and two single-consumer style/constant files (navRowStyles.ts,
contentPadding* in layout/constants.ts) inlined back into their only callers.
