+++ frontend/src/layout/Sidebar.tsx
@@ -22,10 +22,31 @@ import { api } from '@/services/api';
 import { useProfileContext } from '@/contexts/ProfileContext';
 import { useVersion } from '@/contexts/VersionContext';
 import { footerHeight, headerHeight, sidebarWidth } from './constants';
-import { NAV_ROW_SX, navRowTextSlotProps } from './navRowStyles';
 import type { NavItem } from './types';
 import { VersionDisplay } from '@/components/VersionDisplay';
 import { UpdatePanelDialog } from '@/components/UpdatePanelDialog';
+
+// Shared sizing for the sidebar's nav-style rows, kept alongside the one
+// place that uses it — every row is the same height whether or not it
+// happens to carry a subtitle.
+const NAV_ROW_SX = {
+    minHeight: 52,
+    borderRadius: 1.25,
+    py: 1.25,
+    px: 2,
+} as const;
+
+const navRowTextSlotProps = (active: boolean) => ({
+    primary: { noWrap: true, variant: 'body2' as const, sx: { fontWeight: 500, lineHeight: 1.3, fontSize: '0.875rem' } },
+    secondary: {
+        variant: 'caption' as const,
+        sx: {
+            fontSize: '0.6875rem',
+            lineHeight: 1.2,
+            color: active ? 'rgba(255,255,255,0.7)' : 'text.secondary',
+        },
+    },
+});
 
 interface SidebarProps {
     sidebarItems: NavItem[];
