+++ frontend/src/pages/remote-agent/RemoteAgentPage.tsx
@@ -1,28 +1,19 @@
 import { useNavigate, useParams } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
 import { useEffect, useMemo, useState } from 'react';
-import { Telegram, Feishu, Lark, DingTalk, Weixin, WeCom } from '@/components/BrandIcons';
 import { PlatformPicker } from '@/components/bot';
-import { BOT_PLATFORM_IDS, platformDisplayName, usePlatformGuide } from '@/constants/platformGuides';
+import { BOT_PLATFORM_IDS, PLATFORM_BRAND_ICONS, platformDisplayName, usePlatformGuide } from '@/constants/platformGuides';
 import { api } from '@/services/api';
+import { countBotsByPlatform } from '@/types/bot';
 import PlatformRemoteAgentPage from './PlatformRemoteAgentPage';
 
-const PLATFORM_BRAND_ICONS: Record<string, React.ComponentType<{ size?: number; grayscale?: boolean }>> = {
-    telegram: Telegram,
-    feishu: Feishu,
-    lark: Lark,
-    dingtalk: DingTalk,
-    weixin: Weixin,
-    wecom: WeCom,
-};
-
-// RemoteAgentPage is the nav-facing entry for the Remote purpose: ONE sidebar
-// row (under the "Bots" rail icon, alongside Overview and Notify), with
-// platform selection moved in-page — a row of chips above the page content,
-// instead of nine separate sidebar rows. The routes it switches between
-// (/remote-agent/:platform) are unchanged — deep links and the BotCard
-// purpose chip still work. PlatformRemoteAgentPage itself is untouched:
-// same guide, add, and pairing behavior it already had.
+// RemoteAgentPage is the nav-facing entry for the Remote Control purpose:
+// ONE sidebar row (under the "Remote" rail icon, alongside Bots and IM
+// Notify), with platform selection moved in-page — a grid of picker tiles
+// above the page content, instead of nine separate sidebar rows. The routes
+// it switches between (/remote-agent/:platform) are unchanged — deep links
+// and the BotCard purpose chip still work. PlatformRemoteAgentPage itself is
+// untouched: same guide, add, and pairing behavior it already had.
 const RemoteAgentPage = () => {
     const { platform = 'weixin' } = useParams<{ platform: string }>();
     const navigate = useNavigate();
@@ -36,14 +27,7 @@ const RemoteAgentPage = () => {
         let cancelled = false;
         api.getImBotSettingsList().then((data) => {
             if (cancelled || !data?.success || !Array.isArray(data.settings)) return;
-            const map: Record<string, { active: number; total: number }> = {};
-            for (const bot of data.settings) {
-                if (!bot?.platform) continue;
-                const slot = map[bot.platform] ?? (map[bot.platform] = { active: 0, total: 0 });
-                slot.total++;
-                if (bot.enabled) slot.active++;
-            }
-            setCounts(map);
+            setCounts(countBotsByPlatform(data.settings));
         }).catch(() => {});
         return () => { cancelled = true; };
     }, [platform]);
